+++ src/frontend/query_commands.ml
@@ -821,8 +821,12 @@ let dispatch pipeline (type a) : a Query_protocol.t -> a = function
     let cmp l1 l2 = Lexing.compare_pos (loc_start l1) (loc_start l2) in
     (List.sort ~cmp locs, `Not_requested)
   | Inlay_hints
-      (start, stop, hint_let_binding, hint_pattern_binding, avoid_ghost_location)
-    ->
+      ( start,
+        stop,
+        hint_let_binding,
+        hint_pattern_binding,
+        hint_function_params,
+        avoid_ghost_location ) ->
     let start = Mpipeline.get_lexing_pos pipeline start
     and stop = Mpipeline.get_lexing_pos pipeline stop in
     let typer_result = Mpipeline.typer_result pipeline in
@@ -831,7 +835,7 @@ let dispatch pipeline (type a) : a Query_protocol.t -> a = function
       | `Interface _ -> []
       | `Implementation structure ->
         Inlay_hints.of_structure ~hint_let_binding ~hint_pattern_binding
-          ~avoid_ghost_location ~start ~stop structure
+          ~hint_function_params ~avoid_ghost_location ~start ~stop structure
     end
   | Signature_help { position; _ } -> (
     (* Todo: additionnal contextual information could help us provide better