+++ tests/test-dirs/inlay-hint/samples.t
@@ -1,4 +1,139 @@
+Regular function
+
+  $ $MERLIN single inlay-hints -start 1:0 -end 2:26 -avoid-ghost-location false \
+  > -function-params true \
+  > -filename inlay.ml <<EOF
+  > let f a b c d e f = (a + b, c ^ d, e +. (float_of_string f))
+  > EOF
+  {
+    "class": "return",
+    "value": [
+      {
+        "pos": {
+          "line": 1,
+          "col": 17
+        },
+        "label": "string"
+      },
+      {
+        "pos": {
+          "line": 1,
+          "col": 15
+        },
+        "label": "float"
+      },
+      {
+        "pos": {
+          "line": 1,
+          "col": 13
+        },
+        "label": "string"
+      },
+      {
+        "pos": {
+          "line": 1,
+          "col": 11
+        },
+        "label": "string"
+      },
+      {
+        "pos": {
+          "line": 1,
+          "col": 9
+        },
+        "label": "int"
+      },
+      {
+        "pos": {
+          "line": 1,
+          "col": 7
+        },
+        "label": "int"
+      }
+    ],
+    "notifications": []
+  }
+
+Regular function without function-params
+FIXME the option is not enforced
+  $ $MERLIN single inlay-hints -start 1:0 -end 2:26 -avoid-ghost-location false \
+  > -function-params false \
+  > -filename inlay.ml <<EOF
+  > let f a b c d e f = (a + b, c ^ d, e +. (float_of_string f))
+  > EOF
+  {
+    "class": "return",
+    "value": [
+      {
+        "pos": {
+          "line": 1,
+          "col": 17
+        },
+        "label": "string"
+      },
+      {
+        "pos": {
+          "line": 1,
+          "col": 15
+        },
+        "label": "float"
+      },
+      {
+        "pos": {
+          "line": 1,
+          "col": 13
+        },
+        "label": "string"
+      },
+      {
+        "pos": {
+          "line": 1,
+          "col": 11
+        },
+        "label": "string"
+      },
+      {
+        "pos": {
+          "line": 1,
+          "col": 9
+        },
+        "label": "int"
+      },
+      {
+        "pos": {
+          "line": 1,
+          "col": 7
+        },
+        "label": "int"
+      }
+    ],
+    "notifications": []
+  }
+
+
 Optional argument
+
+  $ $MERLIN single inlay-hints -start 1:0 -end 2:26 -avoid-ghost-location false \
+  > -function-params true \
+  > -filename inlay.ml <<EOF
+  > let f ?x () = x ()
+  > EOF
+  {
+    "class": "return",
+    "value": [
+      {
+        "pos": {
+          "line": 1,
+          "col": 8
+        },
+        "label": "'a option"
+      }
+    ],
+    "notifications": []
+  }
+
+Optional argument without function-params
+FIXME the option is not enforced
 
   $ $MERLIN single inlay-hints -start 1:0 -end 2:26 -avoid-ghost-location false \
   > -filename inlay.ml <<EOF
@@ -21,6 +156,7 @@ Optional argument
 Optional argument with value
 
   $ $MERLIN single inlay-hints -start 1:0 -end 2:26 -avoid-ghost-location false \
+  > -function-params true \
   > -filename inlay.ml <<EOF
   > let f ?(x = 1) () = x
   > EOF
@@ -41,6 +177,7 @@ Optional argument with value
 Labeled argument
 
   $ $MERLIN single inlay-hints -start 1:0 -end 2:26 -avoid-ghost-location false \
+  > -function-params true \
   > -filename inlay.ml <<EOF
   > let f ~x = x + 1
   > EOF
@@ -61,6 +198,7 @@ Labeled argument
 Case argument
 
   $ $MERLIN single inlay-hints -start 1:0 -end 2:26 -avoid-ghost-location false \
+  > -function-params true \
   > -filename inlay.ml <<EOF
   > let f (Some x) = x + 1
   > EOF
@@ -81,6 +219,7 @@ Case argument
 Pattern variables without pattern-binding hint
 
   $ $MERLIN single inlay-hints -start 1:0 -end 4:26 -avoid-ghost-location false \
+  > -function-params true \
   > -filename inlay.ml <<EOF
   > let f x =
   >   match x with
@@ -104,7 +243,7 @@ Pattern variables without pattern-binding hint
 Pattern variables with pattern-binding hint
 
   $ $MERLIN single inlay-hints -start 1:0 -end 4:26 -avoid-ghost-location false \
-  > -pattern-binding true \
+  > -pattern-binding true -function-params true \
   > -filename inlay.ml <<EOF
   > let f x =
   >   match x with