+++ CHANGES.md
@@ -4,6 +4,7 @@ unreleased
   + merlin library
     - Expose utilities to manipulate typed-holes in `Merlin_analysis.Typed_hole`
       (#1888)
+    - `inlay-hints` fix inlay hints on function parameters (#1923)
   + vim plugin
     - Added support for search-by-type (#1846)
       This is exposed through the existing `:MerlinSearch` command, that